refactor: 클래스 분리 및 자료구조 개선
